farm2market_backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
Pillow==10.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
